python_backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pypdf
openai
